requirements.txt
pybluez
cryptography